requirements.txt
#
# Core dependencies for Automatic_Cluster_Labeling
#
pandas>=2.0.0
scikit-learn>=1.3.0
datasets>=2.14.0
requests>=2.31.0
sentence-transformers>=3.0.0
networkx>=3.0
torch>=2.0.0
numpy>=2.0.0
groq>=0.9.0
python-dotenv>=1.0.0